ultime aggiunte nei controller e service di Event

+++ src/main/java/java13g3/justMeet/services/EventService.java
@@ -20,14 +20,21 @@ public class EventService {
     }
 
     public void createEvent(Event e) {
-        if (e != null) {
-            
+        if (e.getName().isEmpty() || e.getDescription().isEmpty() || e.getCategory().toString().isEmpty()  ||
+            e.getLanguage().toString().isEmpty()  || e.getCoverPhoto().isEmpty() || e.getEventDate().toString().isEmpty() ||
+            e.getEventApi().isEmpty() || e.getAddressApi().isEmpty() || e.getPrivate().toString().isEmpty() || e.getUserId().isEmpty()){
+            throw new IllegalArgumentException();
+        }else {
             eventRepository.save(e);
         }
     }
 
     public List<Event> retrieveAllEvents() {
-        return eventRepository.findAll();
+        if(eventRepository.findAll().size() >= 1){
+            return eventRepository.findAll();
+        }else {
+            throw new IllegalArgumentException();
+        }
     }
 
     public Optional<Event> retrieveEventById(Long id) {
@@ -97,6 +104,10 @@ public class EventService {
     }
 
     public void deleteEventById(Long id) {
-       eventRepository.deleteById(id);
+        if(eventRepository.findById(id).isPresent()) {
+            eventRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException();
+        }
     }
 }
